+++ test/recursion/recursion.input.ts
@@ -7,3 +7,5 @@ interface Team {
     users: User[]
     parent?: Team
 }
+
+type OneLanguageTranslations = { [key: string]: string | OneLanguageTranslations }